# Conflicts:
#	gradle.properties
#	src/main/kotlin/de/jagenka/User.kt
#	src/main/kotlin/de/jagenka/UserRegistry.kt
#	src/main/resources/fabric.mod.json

+++ src/main/kotlin/de/jagenka/PerformanceMetrics.kt
@@ -1,3 +1,3 @@
 package de.jagenka
 
-data class PerformanceMetrics(val mspt: Double, val tps: Double)
+data class PerformanceMetrics(val mspt: Double, val tps: Float)
